fix: EXPOSED-108 Incorrect mapping for UInt data type (#1809)
Currently, when attempting to insert a UInt value outside of the range [0, 2,147,483,647],
Exposed truncates the value by calling value.toInt() before sending it to the
DB, causing overflow. The value is stored successfully as a negative number because
all databases (except MySQL and MariaDB) don't support unsigned types natively,
which means Exposed is actually mapping to 4-byte INT, which accepts
the range [-2,147,483,648, 2,147,483,647].

Change the default mapping to the next higher-up integer data type BIGINT (technically
a 8-byte storage type) and remove the truncation conversions so that an accurate
value is sent/received to/from the database. To ensure that the intended behavior
cannot be overriden using exec() directly, a check constraint is auto-applied to
the column when registered if the database is not MySQL/MariaDB.

+++ exposed-core/src/main/kotlin/org/jetbrains/exposed/sql/vendors/Default.kt
@@ -39,8 +39,11 @@ abstract class DataTypeProvider {
     /** Numeric type for storing 4-byte integers. */
     open fun integerType(): String = "INT"
 
-    /** Numeric type for storing 4-byte unsigned integers. */
-    open fun uintegerType(): String = "INT"
+    /** Numeric type for storing 4-byte unsigned integers.
+     *
+     * **Note:** If the database being used is not MySQL or MariaDB, this will represent the 8-byte integer type.
+     */
+    open fun uintegerType(): String = "BIGINT"
 
     /** Numeric type for storing 4-byte integers, marked as auto-increment. */
     open fun integerAutoincType(): String = "INT AUTO_INCREMENT"
